# FIX remove lambdas from text preprocessing