Lambda functions are non-serializable under the stdlib pickle
module. This commit replaces the lambdas found in three text
preprocessing functions with hidden functions for chaining
a sequence of preprocessing steps that can be partialed where
appropriate.

#### Reference Issues/PRs

Closes #12833

#### What does this implement/fix? Explain your changes.

Instead of composing functions with lambdas, create chains 
of preprocessing steps inside single functions that can 
be decomposed with partialing.