@thomasjpfan 2019-07-27
Please add an entry to the change log at `doc/whats_new/v0.22.rst`. Like the other entries there, please reference this pull request with `:pr:` and credit yourself (and other contributors if applicable) with `:user:`.