@deniederhut 2019-07-30
Hm... Circle is showing

```
agent key RSA SHA256:oEdX/44m2Y1klY5Uq287qOWuWfJrhusv4e5ImKQ5Uhk returned incorrect signature type
```

Does this need to be rebased?