@rth 2019-07-30
Please resolve conflicts (hopefully that would also fix CI by merging master in).