@deniederhut 2019-08-01
Yup! That did the trick for the CI